Rename all files to camelCase
