+++ packages/cli/test/utils/simulation/clients/validator/index.ts
@@ -1,9 +1,9 @@
 import {writeFile} from "node:fs/promises";
-import {SHARED_JWT_SECRET, SHARED_VALIDATOR_PASSWORD, BN_REST_BASE_PORT} from "../constants.js";
-import {AtLeast, BeaconClient, ValidatorClient, ValidatorGeneratorOptions, ValidatorNode} from "../interfaces.js";
-import {makeUniqueArray} from "../utils/index.js";
-import {createKeystores} from "../utils/keys.js";
-import {ensureDirectories} from "../utils/paths.js";
+import {SHARED_JWT_SECRET, SHARED_VALIDATOR_PASSWORD, BN_REST_BASE_PORT} from "../../constants.js";
+import {AtLeast, BeaconClient, ValidatorClient, ValidatorGeneratorOptions, ValidatorNode} from "../../interfaces.js";
+import {makeUniqueArray} from "../../utils/index.js";
+import {createKeystores} from "../../utils/keys.js";
+import {ensureDirectories} from "../../utils/paths.js";
 import {generateLodestarValidatorNode} from "./lodestar.js";
 import {generateLighthouseValidatorNode} from "./lighthouse.js";
