+++ packages/cli/test/utils/simulation/clients/validator/lighthouse.ts
@@ -4,10 +4,10 @@ import got, {RequestError} from "got";
 import yaml from "js-yaml";
 import {getClient as keyManagerGetClient} from "@lodestar/api/keymanager";
 import {chainConfigToJson} from "@lodestar/config";
-import {RunnerType, ValidatorClient, ValidatorNodeGenerator} from "../interfaces.js";
-import {updateKeystoresPath} from "../utils/keys.js";
-import {getNodeMountedPaths} from "../utils/paths.js";
-import {getNodePorts} from "../utils/ports.js";
+import {RunnerType, ValidatorClient, ValidatorNodeGenerator} from "../../interfaces.js";
+import {updateKeystoresPath} from "../../utils/keys.js";
+import {getNodeMountedPaths} from "../../utils/paths.js";
+import {getNodePorts} from "../../utils/ports.js";
 
 export const generateLighthouseValidatorNode: ValidatorNodeGenerator<ValidatorClient.Lighthouse> = (opts, runner) => {
   if (!process.env.LIGHTHOUSE_BINARY_PATH && !process.env.LIGHTHOUSE_DOCKER_IMAGE) {
